llava/requirements.txt
inflect==7.0.0
nltk==3.8.1
openai==1.10.0
pandas==2.2.0
peft==0.4.0
Pillow==10.2.0
python-dotenv==1.0.1
Requests==2.31.0
rouge_score==0.1.2
tensorflow==2.15.0.post1
torch==2.0.1
transformers==4.36.2
